プロダクトテーブルの追加

+++ src/pages/admin/Products.tsx
@@ -3,6 +3,7 @@ import Select from 'react-select';
 import { Product, Material, MaterialInProduct } from '../../types/types';
 import { getAllProducts, addProduct, updateProduct, deleteProduct } from '../../services/products';
 import { getAllMaterials } from '../../services/materials';
+import { ProductTable } from '../../components/ProductTable';
 
 
 const Products: React.FC = () => {
@@ -13,7 +14,8 @@ const Products: React.FC = () => {
     const [allProducts, setAllProducts] = useState<Product[]>([]);
     const [allMaterials, setAllMaterials] = useState<Material[]>([]);
     const [filteredProducts, setFilteredProducts] = useState<Product[]>([]);
-    const [selectedCategory, setSelectedCategory] = useState<string>('All');  // 選択されたカテゴリ
+    // const [selectedCategory, setSelectedCategory] = useState<string>('All');  // 選択されたカテゴリ
+    const [selectedCategory, setSelectedCategory] = useState<string>('ジン');  // 選択されたカテゴリ
     const [selectedMaterials, setSelectedMaterials] = useState<MaterialInProduct[]>([]);
     const [formProduct, setFormProduct] = useState<Product>({
         id: "",
@@ -39,6 +41,7 @@ const Products: React.FC = () => {
         word: "",
         date: "",
     });
+    const [pendingUpdates, setPendingUpdates] = useState<{ [key: string]: any; id: string }[]>([]);
 
     const categoriesList = ['カクテル', 'ビール', 'ワイン'];  // カテゴリの選択肢
     const baseList = ['ウイスキー', 'ジン', 'ウォッカ', 'ラム'];  // ベースの選択肢
@@ -212,9 +215,16 @@ const Products: React.FC = () => {
     /** 商品更新ボタンを押したときのハンドラ */
     const handleUpdateProducts = async () => {
         console.log("==== 商品の在庫状況と値段を最新化 ====");
+        const isConfirmed = window.confirm('商品の在庫状況と値段を最新化しますか？');
+        if (!isConfirmed) {
+            window.alert('キャンセルしました。');
+            return;
+        }
+
         const allMaterials = await getAllMaterials();
         /** 在庫状況と値段を最新化した Product[] */
         const updatedProducts = allProducts.map(product => {
+            console.log(product.name);
             // 商品に含まれる材料から在庫と値段を最新化
             let isAvailable = true;
             let price = 0;
@@ -239,7 +249,7 @@ const Products: React.FC = () => {
                 return materialInProduct; // 更新後の `materialInProduct` を返す
             });
 
-            price = Math.ceil(price) + 50;  // 氷代 + 消耗品代
+            price = Math.ceil(price) + 40;  // 氷代 + 消耗品代
 
             // 更新対象かどうか確認
             if ((product.isAvailable !== isAvailable) || (product.price !== price)) {
@@ -248,6 +258,7 @@ const Products: React.FC = () => {
                 updateProduct(product.id, {
                     isAvailable: product.isAvailable,
                     price: product.price,
+                    name: product.name,
                 });
                 console.log(`${product.name} の情報をFirestoreに更新しました(` + product.id + ")");
             }
@@ -255,6 +266,7 @@ const Products: React.FC = () => {
         });
 
         setAllProducts(updatedProducts);
+        window.confirm('商品の在庫状況と値段を最新化しました。');
     };
 
     /** カテゴリフィルタの変更ハンドラ */
@@ -270,6 +282,12 @@ const Products: React.FC = () => {
         }
     };
 
+    /** 材料の値を変更したときに変更内容を保持するハンドラ */
+    const handlePendingUpdate = async (updateInfo: { [key: string]: any; id: string; }) => {
+        setPendingUpdates([...pendingUpdates, updateInfo]);
+        console.log(pendingUpdates);
+    };
+
     /** 商品を編集するハンドラ */
     const handleEditProduct = (product: Product) => {
         setFormProduct(product);  // 編集モードで既存データをセット
@@ -278,7 +296,13 @@ const Products: React.FC = () => {
     };
 
     /** 商品を削除するハンドラ */
-    const handleDeleteProduct = async (id: string) => {
+    const handleDeleteRow = async (id: string) => {
+        const isConfirmed = window.confirm('本当に削除しますか？');
+        if (!isConfirmed) {
+            window.alert('キャンセルしました。');
+            return;
+        }
+
         await deleteProduct(id);
         showMessage('商品を削除しました');
         const data = await getAllProducts();
@@ -429,13 +453,19 @@ const Products: React.FC = () => {
                         {filteredProducts.length === 0 ? (
                             <p>該当する商品がありません。</p>
                         ) : (
-                            <ul>
-                                {filteredProducts.map(product => (
-                                    <li key={product.id}>
-                                        <>
-                                            {/* {console.log(product.name)} */}
-                                            {/* {console.log(product.id)} */}
-                                            <h4>{product.name}</h4>
+                            <>
+                                <ProductTable
+                                    products={filteredProducts}
+                                    handlePendingUpdate={handlePendingUpdate}
+                                    handleDeleteRow={handleDeleteRow}
+                                />
+                                {/* <ul> */}
+                                {/* {filteredProducts.map(product => ( */}
+                                {/* <li key={product.id}> */}
+                                {/* <> */}
+                                {/* {console.log(product.name)} */}
+                                {/* {console.log(product.id)} */}
+                                {/* <h4>{product.name}</h4>
                                             値段: ¥ {product.price.toLocaleString()} <br />
                                             カテゴリ: {product.categories.join(', ')} <br />
                                             ベース: {product.bases.join(', ')} <br />
@@ -446,11 +476,12 @@ const Products: React.FC = () => {
                                             レシピ: {product.recipe} <br />
                                             準備完了： {product.already} <br />
                                             <button onClick={() => handleEditProduct(product)}>編集</button>
-                                            <button onClick={() => handleDeleteProduct(product.id)}>削除</button>
-                                        </>
-                                    </li>
-                                ))}
-                            </ul>
+                                            <button onClick={() => handleDeleteRow(product.id)}>削除</button> */}
+                                {/* </> */}
+                                {/* </li> */}
+                                {/* ))} */}
+                                {/* </ul> */}
+                            </>
                         )}
                     </div>
                 )}
